Separate supporting functions into relevant classes
