+++ PolygonFunctions.py
@@ -1,52 +1,13 @@
-""""Supporting functions when trying to work with lines and polygons."""
+""""Supporting functions when working with polygons."""
 
-from shapely import Polygon, LineString, affinity, intersection
-import numpy as np
+from shapely import Polygon, LineString, affinity
 import geopandas
 import matplotlib.pyplot as plt
+import LineFunctions
 
 
 X,Y = 0,1
 linep1idx, linep2idx = 0, 1
-
-
-
-def checkVertical(c1, c2):
-    """Returns True if points are on a vertical line.
-    
-    Vertical lines have different line equations and their normal lines are calculated differently.
-    
-    """
-
-    (x1, y1) = c1
-    (x2, y2) = c2
-    if x2-x1 == 0:
-        return True
-    return False
-
-
-def lineEQ(c1, c2):
-    """Returns line equation for two points as a tuple (gradient, cvalue, isVertical).
-    
-    Line equations are used to find the normal line for two adjacent lines in an rlp,
-    which allows houses to be placed considering both lines so valuable space isn't missed.
-    
-    Line equations are also used to continue to place houses padded from another house,
-    rather than padded from a line, which maximises the rlp area used to place houses.
-    
-    """
-
-    if checkVertical(c1, c2):
-        # should raise an exception here rather than returning a False value.
-        # PADDING WORKS DIFFERENTLY IN THIS CASE. THE NORMAL EQUATION IS JUST SOMETHING ELSE.
-        return (0, 0, True)
-    
-    (x1, y1) = c1
-    (x2, y2) = c2
-    
-    m = (y2-y1)/(x2-x1)
-    c = y1-(m*x1)
-    return (m, c, False)
 
 
 def findLongestLineIndex(polygon):
@@ -96,8 +57,6 @@ def findAdjacentLine(polygon):
                 corner = lcoord
 
     return (lines[adjacentlineindex].coords, corner)
-
-
 
 
 def findAdjacentLines(polygon):
@@ -150,12 +109,12 @@ def findAdjacentLines(polygon):
             adjacentlines.append( LineString( [adjacentpoints[i] , adjacentpoints[i+1]] ) )
             
         
-        # from geopandas import GeoSeries
-        # from shapely import Point
-        # ax=GeoSeries(polygon).plot()
+        from geopandas import GeoSeries
+        from shapely import Point
+        ax=GeoSeries(polygon).plot()
         # GeoSeries( adjacentlines ).plot(ax=ax, color="yellow")
         # GeoSeries( LineString(longestline) ).plot(ax=ax, color="red")
-        # plt.show()
+        plt.show()
         
     elif longestline[linep2idx][X]==maxxvalue:
         """Only use adjacent lines down to the point with the smallest x value"""
@@ -174,9 +133,37 @@ def findAdjacentLines(polygon):
             adjacentlines.append( LineString( [adjacentpoints[i] , adjacentpoints[i+1]] ) )
 
 
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
     else:
         """Have to use adjacent lines in both directions"""
-
+        
 
 
 
@@ -245,9 +232,9 @@ def lines(polygon):
         """Returns the longest line and its adjacent line of a given polygon."""
 
         longestline = findLongestLine(polygon)
-        longestline = orderLine(longestline)
+        longestline = LineFunctions.orderLine(longestline)
         adjacentline, corner = findAdjacentLine(polygon)
-        adjacentline = orderLine(adjacentline)
+        adjacentline = LineFunctions.orderLine(adjacentline)
 
         return longestline, adjacentline, corner
 
@@ -256,50 +243,6 @@ def adjacentlines(polygon):
     pass
 
 
-def orderLine(line):
-    """Returns line as a list. The coordinate with the smaller x-value is first."""
-
-    line = [coord for coord in line]
-
-    if line[linep1idx][X] > line[linep2idx][X]:
-        temp = line[linep1idx]
-        line[linep1idx] = line[linep2idx] 
-        line[linep2idx] = temp 
-
-    return line
-
-
-def findAngle(line):
-    """Returns the angle between two lines, with error from use of pi.
-    
-    Due to pi error, equality comparisons of results of this function should use rounded values.
-
-    """
-
-    xs1, xe1 = line.xy[0]
-    ys1, ye1 = line.xy[1]
-    if xs1 == xe1:
-        angle = np.pi/2
-    else:
-        angle = np.arctan((ye1-ys1)/(xe1-xs1))
-    return angle
-
-def isParallel(line1, line2):
-    """"Returns true if two lines have the same angle to 3 decimal places.
-    
-    This is to account for lines rotated using pi (like with radians).
-
-    """
-
-    angle1 = findAngle(line1)
-    angle2 = findAngle(line2)
-
-    if round(angle1, 3) == round(angle2, 3):
-        return True
-    else:
-        return False
-    
-    
 def rotatePolygon(resultLine, polygon, showRotation=False):
     """"Rotates the polygon until it is parallel to the line.
     
@@ -313,7 +256,7 @@ def rotatePolygon(resultLine, polygon, showRotation=False):
     oldPolygon = polygon
     polygonLine = LineString([polygon.exterior.coords[0], polygon.exterior.coords[1]])
 
-    resultAngle, changeAngle = findAngle(resultLine), findAngle(polygonLine)
+    resultAngle, changeAngle = LineFunctions.findAngle(resultLine), LineFunctions.findAngle(polygonLine)
     polygon = affinity.rotate(polygon, resultAngle-changeAngle, origin="centroid", use_radians=True)
     polygonLine = LineString([polygon.exterior.coords[0], polygon.exterior.coords[1]])
     
@@ -326,8 +269,8 @@ def rotatePolygon(resultLine, polygon, showRotation=False):
 
         print("Original angle of polygonLine is", changeAngle)
         print("The rotated angle of polygonLine should be", resultAngle)
-        print("The actual rotated angle is", findAngle(polygonLine))
-        if isParallel(polygonLine, resultLine):
+        print("The actual rotated angle is", LineFunctions.findAngle(polygonLine))
+        if LineFunctions.isParallel(polygonLine, resultLine):
             print("Successful rotation")
         else:
             print("Unsuccessful rotation")
@@ -359,56 +302,3 @@ def moveToOrigin(polygon, showTranslation=False):
         plt.show()
     
     return Polygon(shiftcoords)
-    
-
-
-
-def leqtoline(leq, polygon):
-        """Returns a line for a line equation. The returned line touches the edges of the polygon."""
-
-        minxval = maxxval = polygon.exterior.coords[0][X]
-        for coord in polygon.exterior.coords:
-            if coord[X]<minxval:
-                minxval=coord[X]
-            if coord[X]>maxxval:
-                maxxval=coord[X]
-
-        verylongline = [(minxval, lineyval(leq, minxval)) , (maxxval, lineyval(leq, maxxval))]
-        shorterline = intersection(polygon, LineString( verylongline ))
-
-        return shorterline
-
-
-def leqs(line):
-    """Returns the line equation for a line."""
-    return lineEQ(line[linep1idx], line[linep2idx])
-
-
-def normalLineEQ(leq, point):
-    """Returns equation for the normal of a given line as a tuple (gradient, cvalue, isVertical).
-    
-    The normal line goes through the given point and the given line, and is used to pad for new rows in the rlp.
-    
-    The given line is given as a tuple (gradient, cvalue, isVertical).
-
-    Returns a special value if the given line is horizontal (returns (0,0,True)).
-    
-    """
-
-    x, y = point[X], point[Y]
-    m, c, isVertical = leq
-
-    if isVertical : return (0, y, False)
-    if m==0 : return (0, 0, True)
-    
-    normalm = -1 / m
-    normalc = y - (normalm*x)
-
-    return (normalm, normalc, False)
-
-
-def lineyval(leq, x):
-    """Returns the y value of a line equation given the x value."""
-    
-    m, c, isVertical = leq
-    return (m*x + c)
